aiml40/requirements.txt
azureml.core
azureml.dataprep